Use this one weird ~trick~ merge policy to prevent duplicates
Every time this app loads, it fetches from the News API to gather more
articles. As an initial measure to prevent duplication and using
out-of-the-box solutions, set a merge policy focused on replacing
what's in persistent store with what's in memory when duplicates are
found.

This happens during the article add process and the constraint is
limited to the `title` property. Naive, but functional for our needs!

Additionally, added a Clear CD button for convenience

Reference:

https://developer.apple.com/documentation/coredata/loading_and_displaying_a_large_data_feed