Implements ability to load articles in WKWebView
SwiftUI currently has no webview, so we use UIViewRepresentable to
integrate a UIKit view (In this case, WKWebView) into our SwiftUI code

The page is loaded via a tap gesture, which pulls up a sheet. From
a functionality perspective, this ain't the best, so I will likely
change it, but it was low barrier way to test this out with what
knowledge I have today!