Adds delete (archive) functionality
Although we are using the onDelete method, the result is changing
the state of the article with `hidden = true`. There's currently
no built-in ability to edit the display of this method.

This helped me understand how I may update an entity from a
FetchedResults set. It was surprisingly simple!

The onDelete method provides our function an IndexSet since you could
easily build a view that allows multiple items to be deleted.

As we iterate over that IndexSet, we can reference the fetched entity
(article) via the provided subscript and simply set values as it were,
nice!

We trust that things will go well and don't implement a do/catch on the
`moc.save()` because .. sample app!