Adds a poorly placed Search bar and allows searches to occur
I didn't want a ton more features to this app but it's a nice place to
add my learnings! I imagine now that I've added this search feature I
might make it easier to manage and search for multiple topics, combining
(and displaying accordingly) that data. That will likely teach me an
additional thing or two about Core Data and data flow in general!